refactor: colocation of related pieces of code
