+++ React Project/src/pages/Login/validationSchema.ts
@@ -1,0 +1,6 @@
+import { object, string } from "yup";
+
+export const loginSchema = object().shape({
+    email: string().email().required("Email is required."),
+    password: string().required("Please enter your password."),
+});
